\name{CUMP-package}
\Rdversion{1.0}
\alias{CUMP-package}
\alias{CUMP}
\docType{package}
\title{
Analyze Multivariate Phenotypes by Combining Univariate results 
}
\description{
This package can be used to detect pleiotropy by combining univariate association test results of multiple 
phenotypes in genome-wide association studies (or studies of a large number of SNPs). In this package, 
we have several combination approaches including the O'Brien's method that is weighted sum of the Z or beta 
statistic (direction sensitive) and other methods that are weighted sum of the squared Z statistics (direction insensitive). 
}
\details{
\tabular{ll}{
Package: \tab CUMP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-04-05\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Qiong Yang <qyang@bu.edu> and Xuan Liu <liuxuan@bu.edu>

Maintainer: Xuan Liu <liuxuan@bu.edu>
}
\references{
CUMP: an R package for analyzing multivariate phenotypes in genetic association studies
}
